Fix renewal of certificates
Zimbra regenerates the patched config files on restart. If certbot_zimbra.sh
is executed on certbot's post-hook, the patch is applied too late and the
renewal process fails. To solve that, certbot_zimbra.sh triggers now cerbot
directly for renewal.

Because of the mentioned changes, executing cerbot_zimbra.sh via cron
should be changed to (path could be different):
/var/lib/certbot-zimbra/certbot_zimbra.sh -r -d $(/opt/zimbra/bin/zmhostname)